requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
langchain_core
groq
langchain-groq
pypdf
stqdm
langchain-openai
google-generativeai